Someone made the intelligent enhancements: 1. allow the passing of access key and secret options. 2. Add a no config file option
We should be able to pass the access key and secret key to the s3cmd script so that they do not need to be stored in the .s3cfg file.
This also makes it easier to jump between accounts through automation scripts without going through the config process for each.
A noconfig option would allow the user to skip the configuration process and settle for the defaults. Of course without further enhancements to the command line options some of the functionality would be missing. However for the basic functionality of ls, put, get, etc. this makes it easy to spin up an EC2 instance, automatically wget s3cmd and immediately start using it without the configuration step.